Added better import resolution for tests, refactored utils into folder
