+++ frontend/jest.config.js
@@ -2,4 +2,6 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  moduleDirectories: ['node_modules', 'src'],
+  rootDir: './src/',
 };
