+++ frontend/src/js/state/epics/metadata.ts
@@ -20,7 +20,7 @@ import {
   show_toast,
 } from '../global';
 
-import { fromWrapper } from './util';
+import { fromWrapper } from './epicUtils';
 import APP_URLS from '../../urls';
 
 import type { MyEpic } from './types';
